Sample .jshintrc file
Template .jshintrc file showing all the options currently explained at http://www.jshint.com/docs/options/

JSHint will start looking for this file in the same directory as the file that's being linted. If not found, it will move one level up the directory tree all the way up to the filesystem root.

All enforcing options are on, all relaxing options are off. Ext global is shown as an example.